refactor: :recycle: Extract bool funcs to check.c
